+++ packages/terradart_google/lib/src/project/terraform_api_requirements.dart
@@ -7,6 +7,7 @@ const List<MapEntry<String, String>> terraformApiPrefixRules = [
   MapEntry('google_gke_hub_', 'gkehub.googleapis.com'),
   MapEntry('google_cloudfunctions2_', 'cloudfunctions.googleapis.com'),
   MapEntry('google_cloud_run_v2_', 'run.googleapis.com'),
+  MapEntry('google_alloydb_', 'alloydb.googleapis.com'),
   MapEntry('google_artifact_registry_', 'artifactregistry.googleapis.com'),
   MapEntry('google_certificate_manager_', 'certificatemanager.googleapis.com'),
   MapEntry('google_privateca_', 'privateca.googleapis.com'),